optional props; wire everything through config
