+++ frontend/src/components/previewers/Preview.tsx
@@ -1,27 +1,9 @@
-import React, { lazy, Suspense, useEffect } from "react";
+import React, { Suspense, useEffect } from "react";
 import { LazyLoadErrorBoundary } from "../errors/LazyLoadErrorBoundary";
 import { useDispatch, useSelector } from "react-redux";
 import { RootState } from "../../types/data";
 import { fetchFileSummary } from "../../actions/file";
-
-const Image = lazy(
-	() => import(/* webpackChunkName: "previewers-image" */ "../previewers/Image")
-);
-const Audio = lazy(
-	() => import(/* webpackChunkName: "previewers-audio" */ "../previewers/Audio")
-);
-const Video = lazy(
-	() => import(/* webpackChunkName: "previewers-video" */ "../previewers/Video")
-);
-
-const Iframe = lazy(
-	() =>
-		import(/* webpackChunkName: "previewers-iframe" */ "../previewers/Iframe")
-);
-
-const Text = lazy(
-	() => import(/* webpackChunkName: "previewers-text" */ "../previewers/Text")
-);
+import { vizConfig } from "../../visualization.config";
 
 type previewProps = {
 	fileId: string;
@@ -42,27 +24,18 @@ export const Preview = (props: previewProps) => {
 	return (
 		<LazyLoadErrorBoundary fallback={<div>Fail to load...</div>}>
 			<Suspense fallback={<div>Loading...</div>}>
-				{fileSummary.content_type && (
-					<>
-						{fileSummary.content_type.main_type === "image" && (
-							<Image fileId={fileId} />
-						)}
-						{fileSummary.content_type.main_type === "audio" && (
-							<Audio fileId={fileId} />
-						)}
-						{fileSummary.content_type.main_type === "video" && (
-							<Video fileId={fileId} />
-						)}
-						{fileSummary.content_type.main_type === "text" &&
-							fileSummary.content_type.content_type === "text/html" && (
-								<Iframe fileId={fileId} />
-							)}
-						{fileSummary.content_type.main_type === "text" &&
-							fileSummary.content_type.content_type !== "text/html" && (
-								<Text fileId={fileId} />
-							)}
-					</>
-				)}
+				{(() => {
+					return Object.keys(vizConfig).map((type) => {
+						if (
+							fileSummary &&
+							fileSummary.content_type !== undefined &&
+							type === fileSummary.content_type.main_type
+						) {
+							return React.cloneElement(vizConfig[type], { fileId: fileId });
+						}
+						return null; // Handle the case when the content_type doesn't match any vizConfig type
+					});
+				})()}
 			</Suspense>
 		</LazyLoadErrorBoundary>
 	);
